requirements.txt
path
thefuzz
pymupdf
openai
python-dotenv
fuzzysearch